+++ test/ex_xendit/xen_platform_test.exs
@@ -27,4 +27,20 @@ defmodule ExXendit.XenPlatformTest do
       end
     end
   end
+
+  describe "create_transfer/1" do
+    test "will return result" do
+      use_cassette "valid_create_transfer" do
+        params = %{
+          reference: "test_a2",
+          amount: 60,
+          source_user_id: "63d23daedfa237a41007760b",
+          destination_user_id: "655438b31991c5156ae7a348"
+        }
+
+        assert {:ok, %{body: body}} = XenPlatform.create_transfer(params)
+        assert %{"status" => "SUCCESSFUL"} = body
+      end
+    end
+  end
 end